requirements.txt
Faker
Flask
Flask-Bcrypt
Flask-JWT-Extended
dlib
numpy
opencv-python
pandas
pillow
python-dotenv
pymongo
shutil
